+++ src/app/components/categories-page/category-create/category-search/category-search.ts
@@ -8,10 +8,10 @@ import {
   OnInit,
   viewChild,
 } from '@angular/core';
-import { SearchService } from '../../../../core/services/search-service';
+import { SearchService } from '../../../../core/services/search-service/search-service';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Subject, switchMap } from 'rxjs';
-import { SearchStateService } from '../../../../core/services/search-state-service';
+import { SearchStateService } from '../../../../core/services/search-state-service/search-state-service';
 import { ITunesTrack } from '../../../../models/itunes.model';
 import { FormsModule } from '@angular/forms';
 import { Ellipsis } from '../../../../shared/directives/ellipsis/ellipsis';
