+++ src/app/components/forms/register-form/register-form.ts
@@ -9,7 +9,7 @@ import {
   viewChild,
 } from '@angular/core';
 import { FormControl, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
-import { AuthService } from '../../../core/services/auth-service';
+import { AuthService } from '../../../core/services/auth-service/auth-service';
 import { firebasePasswordValidator, namePatternValidator } from '../../../shared/utils/validators';
 import { getErrorMessage } from '../../../shared/utils/get-error-message';
 import { getAuthError } from '../../../shared/utils/get-auth-error';
