+++ src/app/core/services/search-service/search-service.ts
@@ -1,7 +1,7 @@
 import { inject, Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { map, Observable } from 'rxjs';
-import { ITunesResponse, ITunesTrack } from '../../models/itunes.model';
+import { ITunesResponse, ITunesTrack } from '../../../models/itunes.model';
 
 @Injectable({
   providedIn: 'root',
