+++ src/app/core/services/search-state-service/search-state-service.ts
@@ -1,5 +1,5 @@
 import { Injectable, signal } from '@angular/core';
-import { ITunesTrack } from '../../models/itunes.model';
+import { ITunesTrack } from '../../../models/itunes.model';
 
 @Injectable({
   providedIn: 'root',
