+++ src/app/layout/user-menu/user-menu.ts
@@ -1,5 +1,5 @@
 import { Component, inject } from '@angular/core';
-import { AuthService } from '../../core/services/auth-service';
+import { AuthService } from '../../core/services/auth-service/auth-service';
 import { Router, RouterModule } from '@angular/router';
 import { TranslatePipe } from '@ngx-translate/core';
