+++ src/app/shared/directives/result-color/result-color.spec.ts
