+++ src/app/shared/directives/result-color/result-color.ts
