+++ src/app/shared/directives/theme-icon/theme-icon.spec.ts
